+++ 1.21.11/Common/src/main/java/fuzs/universalenchants/init/ModRegistry.java
@@ -3,33 +3,33 @@ package fuzs.universalenchants.init;
 import fuzs.puzzleslib.api.data.v2.AbstractDatapackRegistriesProvider;
 import fuzs.puzzleslib.api.init.v3.tags.TagFactory;
 import fuzs.universalenchants.UniversalEnchants;
-import net.minecraft.advancements.criterion.DamageSourcePredicate;
-import net.minecraft.advancements.criterion.TagPredicate;
+import net.minecraft.advancements.criterion.*;
 import net.minecraft.core.HolderGetter;
+import net.minecraft.core.HolderSet;
 import net.minecraft.core.RegistrySetBuilder;
 import net.minecraft.core.Vec3i;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.sounds.SoundEvents;
 import net.minecraft.tags.*;
+import net.minecraft.util.valueproviders.ConstantFloat;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlotGroup;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.enchantment.Enchantment;
-import net.minecraft.world.item.enchantment.EnchantmentEffectComponents;
-import net.minecraft.world.item.enchantment.Enchantments;
-import net.minecraft.world.item.enchantment.LevelBasedValue;
-import net.minecraft.world.item.enchantment.effects.AddValue;
-import net.minecraft.world.item.enchantment.effects.DamageImmunity;
-import net.minecraft.world.item.enchantment.effects.ReplaceDisk;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.enchantment.*;
+import net.minecraft.world.item.enchantment.effects.*;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
 import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProvider;
 import net.minecraft.world.level.material.Fluids;
-import net.minecraft.world.level.storage.loot.predicates.DamageSourceCondition;
-import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceCondition;
+import net.minecraft.world.level.storage.loot.LootContext;
+import net.minecraft.world.level.storage.loot.predicates.*;
 
+import java.util.List;
 import java.util.Optional;
 
 public class ModRegistry {
@@ -51,14 +51,20 @@ public class ModRegistry {
         return TagKey.create(Registries.ITEM, resourceKey.identifier().withPrefix("secondary_enchantable/"));
     }
 
+    public static TagKey<Enchantment> getExclusiveSetEnchantmentTag(ResourceKey<Enchantment> resourceKey) {
+        return TagKey.create(Registries.ENCHANTMENT, resourceKey.identifier().withPrefix("exclusive_set/"));
+    }
+
     public static TagKey<Enchantment> getInclusiveSetEnchantmentTag(ResourceKey<Enchantment> resourceKey) {
         return TagKey.create(Registries.ENCHANTMENT, resourceKey.identifier().withPrefix("inclusive_set/"));
     }
 
     public static void boostrapEnchantments(BootstrapContext<Enchantment> context) {
+        HolderGetter<Block> blockLookup = context.lookup(Registries.BLOCK);
         HolderGetter<Item> itemLookup = context.lookup(Registries.ITEM);
+        HolderGetter<EntityType<?>> entityLookup = context.lookup(Registries.ENTITY_TYPE);
         HolderGetter<Enchantment> enchantmentLookup = context.lookup(Registries.ENCHANTMENT);
-        // allow frost walker to replace sea vegetation and itself, also remove on ground check to enable jump-sprinting across water
+        // Allow frost walker to replace sea vegetation and itself, also remove on ground check to enable jump-sprinting across water.
         ReplaceDisk replaceDisk = new ReplaceDisk(new LevelBasedValue.Clamped(LevelBasedValue.perLevel(3.0F, 1.0F),
                 0.0F,
                 16.0F),
@@ -90,7 +96,7 @@ public class ModRegistry {
                         .withEffect(EnchantmentEffectComponents.TICK, replaceDisk,
                                 // has a chance of about 90% to tick at least once every second, which should be enough
                                 LootItemRandomChanceCondition.randomChance(0.1F)));
-        // remove arrow entity type check, so this also works for tridents
+        // Remove the arrow entity type check, so this also works for tridents.
         AbstractDatapackRegistriesProvider.registerEnchantment(context,
                 Enchantments.POWER,
                 Enchantment.enchantment(Enchantment.definition(itemLookup.getOrThrow(ItemTags.BOW_ENCHANTABLE),
@@ -101,5 +107,52 @@ public class ModRegistry {
                                 1,
                                 EquipmentSlotGroup.MAINHAND))
                         .withEffect(EnchantmentEffectComponents.DAMAGE, new AddValue(LevelBasedValue.perLevel(0.5F))));
+        // Allow entities attacking with a mace. Must be a smash attack; that is copied from the Wind Burst enchantment.
+        AbstractDatapackRegistriesProvider.registerEnchantment(context,
+                Enchantments.CHANNELING,
+                Enchantment.enchantment(Enchantment.definition(itemLookup.getOrThrow(ItemTags.TRIDENT_ENCHANTABLE),
+                                1,
+                                1,
+                                Enchantment.constantCost(25),
+                                Enchantment.constantCost(50),
+                                8,
+                                EquipmentSlotGroup.MAINHAND))
+                        .withEffect(EnchantmentEffectComponents.POST_ATTACK,
+                                EnchantmentTarget.ATTACKER,
+                                EnchantmentTarget.VICTIM,
+                                AllOf.entityEffects(new SummonEntityEffect(HolderSet.direct(EntityType.LIGHTNING_BOLT.builtInRegistryHolder()),
+                                                false),
+                                        new PlaySoundEffect(List.of(SoundEvents.TRIDENT_THUNDER),
+                                                ConstantFloat.of(5.0F),
+                                                ConstantFloat.of(1.0F))),
+                                AllOfCondition.allOf(WeatherCheck.weather().setThundering(true),
+                                        LootItemEntityPropertyCondition.hasProperties(LootContext.EntityTarget.THIS,
+                                                EntityPredicate.Builder.entity()
+                                                        .located(LocationPredicate.Builder.location()
+                                                                .setCanSeeSky(true))),
+                                        AnyOfCondition.anyOf(LootItemEntityPropertyCondition.hasProperties(LootContext.EntityTarget.DIRECT_ATTACKER,
+                                                        EntityPredicate.Builder.entity().of(entityLookup, EntityType.TRIDENT)),
+                                                LootItemEntityPropertyCondition.hasProperties(LootContext.EntityTarget.DIRECT_ATTACKER,
+                                                        EntityPredicate.Builder.entity()
+                                                                .equipment(EntityEquipmentPredicate.Builder.equipment()
+                                                                        .mainhand(ItemPredicate.Builder.item()
+                                                                                .of(itemLookup, Items.MACE)))
+                                                                .flags(EntityFlagsPredicate.Builder.flags()
+                                                                        .setIsFlying(false))
+                                                                .moving(MovementPredicate.fallDistance(MinMaxBounds.Doubles.atLeast(
+                                                                        1.5)))))))
+                        .withEffect(EnchantmentEffectComponents.HIT_BLOCK,
+                                AllOf.entityEffects(new SummonEntityEffect(HolderSet.direct(EntityType.LIGHTNING_BOLT.builtInRegistryHolder()),
+                                                false),
+                                        new PlaySoundEffect(List.of(SoundEvents.TRIDENT_THUNDER),
+                                                ConstantFloat.of(5.0F),
+                                                ConstantFloat.of(1.0F))),
+                                AllOfCondition.allOf(WeatherCheck.weather().setThundering(true),
+                                        LootItemEntityPropertyCondition.hasProperties(LootContext.EntityTarget.THIS,
+                                                EntityPredicate.Builder.entity().of(entityLookup, EntityType.TRIDENT)),
+                                        LocationCheck.checkLocation(LocationPredicate.Builder.location()
+                                                .setCanSeeSky(true)
+                                                .setBlock(net.minecraft.advancements.criterion.BlockPredicate.Builder.block()
+                                                        .of(blockLookup, BlockTags.LIGHTNING_RODS))))));
     }
 }