+++ 1.21.11/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.11.2-1.21.11] - 2025-12-26
+
+### Added
+
+- Add Channeling to the mace, which takes effect when striking with a smash attack
+- This makes Channeling incompatible with the Wind Burst enchantment
+
 ## [v21.11.1-1.21.11] - 2025-12-25
 
 ### Changed